requirements.txt
opencv-python
numpy
pythreejs
flet
trimesh
Pillow
pyglet
